+++ apps/workflows-dashboard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/workflows-dashboard
 
+## 0.2.0
+
+### Minor Changes
+
+- version bump
+
 ## 0.1.15
 
 ### Patch Changes